number of times the product is orderd,check quantity ,dec. quantity available,add and remove role on susbend and activate

+++ src/main/java/onlineStore/plantsStore/orders/orderService.java
@@ -1,6 +1,7 @@
 package onlineStore.plantsStore.orders;
 
 import onlineStore.plantsStore.cart.cart;
+import onlineStore.plantsStore.products.product;
 import onlineStore.plantsStore.users.users;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
@@ -14,22 +15,52 @@ public class orderService {
     private final onlineStore.plantsStore.cart.cartRepository cartRepository;
     private final orderRepository orderRepository;
     private final onlineStore.plantsStore.users.usersRepository usersRepository;
+    private final onlineStore.plantsStore.products.productRepository productRepository;
     @Autowired
-    public orderService(onlineStore.plantsStore.cart.cartRepository cartRepository, orderRepository orderRepository, onlineStore.plantsStore.users.usersRepository usersRepository){
+    public orderService(onlineStore.plantsStore.cart.cartRepository cartRepository, orderRepository orderRepository, onlineStore.plantsStore.users.usersRepository usersRepository, onlineStore.plantsStore.products.productRepository productRepository){
         this.cartRepository = cartRepository;
         this.orderRepository = orderRepository;
         this.usersRepository = usersRepository;
+        this.productRepository = productRepository;
     }
 
     public  void order(String username){
 
         List<cart> cartItems= new ArrayList<>();
         cartItems = cartRepository.findcartByUser(username);
+        if(!isQuantityAvalable(cartItems)){
+            throw new IllegalStateException("There Is No Available Quantity");
+        }
         users u = usersRepository.findusersByEmail(username);
         List<orders> orders = cartToOrder(cartItems,u.NewOrderCount());
         orderRepository.saveAll(orders);
+        decAvalable( cartItems);
+        incProductNum(cartItems);
         usersRepository.save(u);
         cartRepository.deleteAll(cartItems);
+    }
+    public boolean isQuantityAvalable(List<cart> cartItems){
+        for (int i = 0; i < cartItems.size(); i++) {
+            product p = productRepository.getById(cartItems.get(i).getId().getProductID());
+            if(p.getQuantityAvailable()<cartItems.get(i).getQuantity()){
+                return false;
+            }
+        }
+        return true;
+    }
+    public void decAvalable(List<cart> cartItems){
+        for (int i = 0; i < cartItems.size(); i++) {
+            product p = productRepository.getById(cartItems.get(i).getId().getProductID());
+            p.decquantityAvailable(cartItems.get(i).getQuantity());
+            productRepository.save(p);
+        }
+    }
+    public void incProductNum(List<cart> cartItems){
+        for (int i = 0; i < cartItems.size(); i++) {
+            product p = productRepository.getById(cartItems.get(i).getId().getProductID());
+            p.incNumOrderd(cartItems.get(i).getQuantity());
+            productRepository.save(p);
+        }
     }
     public List<orders> cartToOrder(List<cart> cartList,int orderNo){
         List<orders> orders = new ArrayList<>();
